Fix MSVC warning
ksycocadict.cpp(562): warning C4267: 'initializing': conversion from
'size_t' to 'const qint32', possible loss of data

QIODevice's offset (passed in seek()) is a qint64 anyway.

+++ src/sycoca/ksycocadict.cpp
@@ -559,7 +559,7 @@ qint32 KSycocaDictPrivate::offsetForKey(const QString &key) const
     const uint hash = hashKey(key) % hashTableSize;
     //qCDebug(SYCOCA) << "hash is" << hash;
 
-    const qint32 off = offset + sizeof(qint32) * hash;
+    const qint64 off = offset + sizeof(qint32) * hash;
     //qCDebug(SYCOCA) << QString("off is %1").arg(off,8,16);
     stream->device()->seek(off);
